refactor(spatial-ai): extract segmentation cache helpers
